slider alignment

+++ trunk/cpsycle/host/src/pianoroll.h
@@ -86,9 +86,9 @@ INLINE void pianokeyboardstate_clip(KeyboardState* self,
 {
 	assert(self);
 
-	*rv_keymin = psy_max(self->keymin, (uint8_t)
+	*rv_keymin = (uint8_t)psy_max((int16_t)self->keymin, (int16_t)
 		((self->keyboardheightpx - clip_bottom_px) / self->keyheightpx - 1));
-	*rv_keymax = psy_min(self->keymax, (uint8_t)
+	*rv_keymax = (uint8_t)psy_min((int16_t)self->keymax, (int16_t)
 		((self->keyboardheightpx - clip_top_px + self->keyheightpx) /
 		self->keyheightpx));
 }
